Fix ARM CI with wrong usage of "arm64-graviton2" resources
In # apache/storm/pull/3344, we have enabled ARM CI with all Storm
modules enabled and switched to use the "arm64-graviton2" resources, but
I have made a mistak that the ARM CI actually still run on x86 than
ARM even the CI job shows the label "arm64-graviton2", details can be
found in the CI job logs[1]. Because the "arm64-graviton2" requires
explicit virt: [lxd|vm] tag, see[2].

Addtionally, we have skip the increasing swap space statement in
intergration test because it will raise error on ARM. And we only
enabled openjdk11 for ARM jobs because the jdk8 is unsupported, see[3].

[1] https://travis-ci.org/github/apache/storm/jobs/741780900
[2] https://docs.travis-ci.com/user/multi-cpu-architectures#testing-on-multiple-cpu-architectures
[3] https://travis-ci.com/github/liusheng/storm/jobs/430934514

+++ integration-test/run-it.sh
@@ -36,10 +36,13 @@ then
   chmod o+rx /home/travis
 fi
 list_storm_processes || true
-# increasing swap space so we can run lots of workers
-sudo dd if=/dev/zero of=/swapfile.img bs=4096 count=1M
-sudo mkswap /swapfile.img
-sudo swapon /swapfile.img
+if [ "$(uname -m)" != aarch64 ]; then
+  # increasing swap space so we can run lots of workers
+  sudo dd if=/dev/zero of=/swapfile.img bs=4096 count=1M
+  sudo mkswap /swapfile.img
+  sudo swapon /swapfile.img
+fi
+
 
 if [[ "${USER}" == "vagrant" ]]; then # install oracle jdk8 or openjdk11
     sudo apt-get update
